requirements.txt
Django>=4.0,<5.0
pandas>=1.5,<2.1
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<2.10
gunicorn>=20.0,<21.0
whitenoise>=6.0,<7.0
django-environ>=0.9,<0.10
numpy>=1.23,<1.27
opencv-python>=4.5,<4.10
tensorflow>=2.10,<2.17
scikit-learn>=1.0,<1.5
Pillow>=9.0,<10.2
pytz
sqlparse